case 4 anomaly detection by forecast/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
torch>=2.0.0
timesfm[torch] @ git+https://github.com/google-research/timesfm.git
python-dotenv>=1.0.0
